SOLR-XXXXX: Add "terms" postfilter support
Prior to this commit, the "terms" query had no support for running as a
postfilter.  This commit adds a DocValuesTermsQuery subclass which does
support postfilter-mode.

The postfilter implementation uses a composite SortedSetDocValues object
to create a doc-values bitset where only the terms on the "terms" query
are set.  At post-filter time the TermsCollector class advances the
SortedSetDocValues object to keep up with the doc being collected, only
collecting each doc if it has a doc-value ord that is set in the bitset.