Reproduce missed JdbcDataSource closure
This commit adds a bit of logging and a bash script to reproduce
a resource-management bug in DIH.

When a DIH run hits IO errors sending documents to other nodes, DIH
doesn't properly clean up its resources.  In practice this leaks
db connections and other resources that can add up over multiple DIH
runs.

To try this out locally, move to the 'solr' directory and run:
    './reproduce-dih-jdbc-leak.sh'

This will spin up two Solr nodes configured for DIH and kick off a DIH
run.  This commit has mocked up ConcurrentUpdateHttp2SolrClient to throw
a fake IOException, which reproduces the problem.  After DIH has a
chance to finish, grep the logs for "In JdbcDataSource.close()" which is
printed as soon as that method is entered.  It should not be present,
proving the problem.

To see how this behaves in the "normal" case edit the
reproduce-dih-jdbc-leak.sh script, changing the "CAUSE_DIH_IO_ERROR"
variable to false.  This will allow DIH to run without the fake IO
issues, and the logs will contain several "In JdbcDataSource.close()"
messages.