LUCENESOLR-1983: CP-79: GCS Backup Repository
This commit adds the GCS Backup Repository as a new contrib to Solr. There are test issues e.g. in the backup-managed-solr bucket, the test expects a pre-created /backup1 folder that need to be created. Also, the version dependencies are not present in the lock file which needs to be fixed. It also adds four new rules to the test security policy, one of which is platform specific (it loads google's credentials from the user's home directory) and has never been tested on windows.