Make BackupId more of a value-type
BackupId doesn't just wrap the ID, it also has a lot of methods for
converting that ID into a filename, etc.  This commit moves someo f
these to 'BackupFilePaths' (a rename of IncrementalFilePaths to be
generic to both incremental and traditional backup).

There are still some BackupId methods that need moved, particularly
'findAll', and 'findMostRecent'.  But the class needs moved to a
different package and tests need migrated, so I'm drawing the line here.